requirements.txt
langgraph>=0.2.50
langchain-anthropic>=0.3.0
langchain-core>=0.3.20
langchain-mcp-adapters>=0.1.0
anthropic>=0.40.0
pydantic>=2.9
python-dotenv>=1.0
cairosvg>=2.7
